Add files via upload
## Power BI Dashboard for Pizza Sales Insights

This file contains a Power BI dashboard that visually represents the insights derived from the pizza sales data. The interactive dashboard showcases various metrics and trends, making it easy to understand the analysis at a glance. It provides a dynamic way to explore the data visually.

To Use:
1. Download and install Power BI Desktop.
2. Open this file in Power BI Desktop.
3. Interact with the dashboard to explore metrics and insights.

Data Source: 'pizza_sales.csv'

-- Tausif Perwez